Refactor: 모듈화 (service, view, handler)
